compiling where clauses

+++ tessera/internal/expression.cpp
@@ -129,7 +129,7 @@ tess::expr_value tess::addition_expr::eval( tess::evaluation_context& ctxt) cons
 
 void tess::addition_expr::compile(stack_machine::stack& stack) const
 {
-	stack.push(std::make_shared<add_op>(terms_.size()));
+	stack.push(std::make_shared<add_op>( static_cast<int>(terms_.size()) ));
 	for (auto [sign, term_expr] : terms_) {
 		term_expr->compile(stack);
 		if (!sign)
